Worker Deployment API (#798)
* Update Infra code structure

* Update script with pygrid_0.3

* Added worker api routes

* Inlcuded basic worker api skeleton routes
* Added Worker db.Model
* Modified Provider class to support destorying deployments

* Initial infra module inside domain

* `Provider` super class

* `Terraform` class

* Move aws provider

* Work on GCP provider

##TODO: need to be completed and TESTED

* Modified get_instance_type() to include categorization of instances

* fetches various instance categories and corresponding instance types for the user

* Added Acclerated computing instances to instance types

* Fix /workers and /workers/<int:id> routes

* Added workers delete route

* Refactored Terraform and Provider classes

* Backward compatiblity bug fix

* Added states in workers deployment

* moved worker-api to domain folder

* Refactor AWS_Serverfull class to support worker deployment

* infra latest work

* Read worker config from domain env variables

* write env vars after deploying domain

* Update TF

* Add Domain DB model

* Update CLI

* Update exec_script in serverfull

* Update Provider

* Update Infra API App

* Use domain credential to deploy worker

* Support pre-built loadings of plugins on terraform init

* Added general purpose instances to cli instance list

* removed domain/apps/core/infra

* Remove infra files from app/domain/src

* Add test route to check domain deployment

* Added suggested changes

* Fix all deployment issues & write env vars to file

* Fix bugs

* Removed apps/domain/src/main/infra

* Moved worker-api to apps/src/domain

* Fix poetry issues

* Remove dummy route

* Fix bugs

* Copy apps/infrastructure to apps/domain/core/infrastructure

* Added terrascript as a dependency

* Add PyInquirer as a dependency

* Write credentails on domain after deployment

* log cred_file

* use jq to write credentials.json

* Install terraform on domain

* Use normal file instead of json for credentials on domain

* bug fix

* Read credentials from env vars

* Merged workers db model to environments db model

* Integrate worker-api routes to  infra_services

* Add admin rights to get_worker_msg

* Fix delete worker infra_service

* Fix bugs

* Fix bugs

* Fix bugs in deleting workers

* Fix worker deletion and add POST method to get_all_workers

* Fix bugs

* Fix bugs in get_workers

* Add url params to get_workers

* Fix bugs in get_all_workers route

* Deleted apps/infrastructure/worker_api/

* remove test route

* Remove dev helpers code

* Fix Response type

* Fix code style

* Update Environment CI tests

* Ignore Integration infrastructure integration tests

* Run docformatter

+++ apps/infrastructure/providers/aws/aws.py
@@ -1,3 +1,4 @@
+# from ..terraform import generate_cidr_block, var, var_module
 from ...tf import generate_cidr_block, var, var_module
 from ..provider import *
 
@@ -5,34 +6,27 @@ from ..provider import *
 class AWS(Provider):
     """Amazon Web Services (AWS) Cloud Provider."""
 
-    def __init__(self, config: Config) -> None:
+    def __init__(self, config: SimpleNamespace) -> None:
         """
-        config (Config) : Object storing the required configuration for deployment
+        config (SimpleNamespace) : Object storing the required configuration for deployment
         """
-        super().__init__(app=config.app.name)
+        super().__init__(config)
         self.config = config
 
         credentials_dir = os.path.join(str(Path.home()), ".aws/api/")
         os.makedirs(credentials_dir, exist_ok=True)
-        self.cred_file = os.path.join(credentials_dir, "credentialss.json")
+        self.cred_file = os.path.join(credentials_dir, "credentials.json")
 
-        with open(self.cred_file, "w") as f:
-            json.dump(vars(config.credentials.cloud), f, indent=2, sort_keys=False)
-
-        self.region = config.vpc.region
-        self.av_zones = config.vpc.av_zones
+        if not os.path.exists(self.cred_file):
+            with open(self.cred_file, "w") as f:
+                json.dump(vars(config.credentials.cloud), f, indent=2, sort_keys=False)
 
         self.tfscript += terrascript.provider.aws(
-            region=self.region, shared_credentials_file=self.cred_file
+            region=self.config.vpc.region, shared_credentials_file=self.cred_file
         )
 
-        # Build the Infrastructure
         self.vpc = None
         self.subnets = []
-        self.build_vpc()
-        self.build_igw()
-        self.build_public_rt()
-        self.build_subnets()
 
     def build_vpc(self):
         """Adds a VPC."""
@@ -90,7 +84,7 @@ class AWS(Provider):
         - one Route table : Routes the traffic from the NAT gateway to the private subnet
         """
 
-        for i, av_zone in enumerate(self.av_zones):
+        for i, av_zone in enumerate(self.config.vpc.av_zones):
             private_subnet = resource.aws_subnet(
                 f"private-subnet-{i}",
                 vpc_id=var(self.vpc.id),
@@ -98,6 +92,7 @@ class AWS(Provider):
                     base_cidr_block=self.vpc.cidr_block, netnum=(2 * i)
                 ),
                 availability_zone=av_zone,
+                map_public_ip_on_launch=True,
                 tags={"Name": f"private-{i}"},
             )
             self.tfscript += private_subnet
@@ -109,6 +104,7 @@ class AWS(Provider):
                     base_cidr_block=self.vpc.cidr_block, netnum=(2 * i + 1)
                 ),
                 availability_zone=av_zone,
+                map_public_ip_on_launch=True,
                 tags={"Name": f"public-{i}"},
             )
             self.tfscript += public_subnet
